var_api/requirements.txt
# ===== requirements.txt =====

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
open-clip-torch>=2.20.0